refactory(#14): 엔티티 수정
투표기능으로 수정된 erd 반영
