+++ src/main/java/com/prography/yakgwa/domain/meet/controller/res/MeetWithStatusInfoResponse.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.prography.yakgwa.domain.meet.entity.Meet;
 import com.prography.yakgwa.domain.meet.entity.MeetStatus;
 import com.prography.yakgwa.domain.meet.service.req.MeetWithVoteAndStatus;
-import com.prography.yakgwa.domain.vote.entity.PlaceVote;
-import com.prography.yakgwa.domain.vote.entity.TimeVote;
+import com.prography.yakgwa.domain.vote.entity.place.PlaceVote;
+import com.prography.yakgwa.domain.vote.entity.time.TimeVote;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
