+++ src/main/java/com/prography/yakgwa/domain/meet/service/req/MeetWithVoteAndStatus.java
@@ -2,8 +2,8 @@ package com.prography.yakgwa.domain.meet.service.req;
 
 import com.prography.yakgwa.domain.meet.entity.Meet;
 import com.prography.yakgwa.domain.meet.entity.MeetStatus;
-import com.prography.yakgwa.domain.vote.entity.PlaceVote;
-import com.prography.yakgwa.domain.vote.entity.TimeVote;
+import com.prography.yakgwa.domain.vote.entity.place.PlaceVote;
+import com.prography.yakgwa.domain.vote.entity.time.TimeVote;
 import lombok.Builder;
 import lombok.Getter;
