+++ src/main/java/com/prography/yakgwa/domain/vote/entity/place/PlaceSlot.java
@@ -1,0 +1,21 @@
+package com.prography.yakgwa.domain.vote.entity.place;
+
+import com.prography.yakgwa.domain.meet.entity.Meet;
+import com.prography.yakgwa.domain.place.entity.Place;
+import com.prography.yakgwa.global.entity.BaseTimeEntity;
+import jakarta.persistence.*;
+
+@Entity(name = "PLACESLOT_TABLE")
+public class PlaceSlot extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private boolean confirm;
+    @ManyToOne
+    @JoinColumn(name = "meet_id")
+    private Meet meet;
+
+    @ManyToOne
+    @JoinColumn(name = "place_id")
+    private Place place;
+}
