+++ src/main/java/com/prography/yakgwa/domain/vote/entity/time/TimeVote.java
@@ -1,4 +1,4 @@
-package com.prography.yakgwa.domain.vote.entity;
+package com.prography.yakgwa.domain.vote.entity.time;
 
 import com.prography.yakgwa.domain.meet.entity.Meet;
 import com.prography.yakgwa.domain.user.entity.User;
@@ -19,13 +19,10 @@ public class TimeVote {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private LocalDateTime time;
-    @Builder.Default
-    private Boolean confirm=Boolean.FALSE;
     @ManyToOne
     @JoinColumn(name = "user_id")
     private User user;
     @ManyToOne
-    @JoinColumn(name = "meet_id")
-    private Meet meet;
+    @JoinColumn(name = "timeslot_id")
+    private TimeSlot timeSlot;
 }
