+++ src/main/java/com/prography/yakgwa/domain/vote/impl/TimeVoteReader.java
@@ -1,6 +1,6 @@
 package com.prography.yakgwa.domain.vote.impl;
 
-import com.prography.yakgwa.domain.vote.entity.TimeVote;
+import com.prography.yakgwa.domain.vote.entity.time.TimeVote;
 import com.prography.yakgwa.domain.vote.repository.TimeVoteJpaRepository;
 import com.prography.yakgwa.global.meta.ImplService;
 import lombok.RequiredArgsConstructor;
