+++ src/main/java/com/prography/yakgwa/domain/vote/impl/dto/VoteInfoWithStatus.java
@@ -1,8 +1,8 @@
 package com.prography.yakgwa.domain.vote.impl.dto;
 
 import com.prography.yakgwa.domain.meet.entity.MeetStatus;
-import com.prography.yakgwa.domain.vote.entity.PlaceVote;
-import com.prography.yakgwa.domain.vote.entity.TimeVote;
+import com.prography.yakgwa.domain.vote.entity.place.PlaceVote;
+import com.prography.yakgwa.domain.vote.entity.time.TimeVote;
 import lombok.Builder;
 import lombok.Getter;
