+++ src/main/java/com/prography/yakgwa/domain/vote/repository/PlaceVoteJpaRepository.java
@@ -1,6 +1,6 @@
 package com.prography.yakgwa.domain.vote.repository;
 
-import com.prography.yakgwa.domain.vote.entity.PlaceVote;
+import com.prography.yakgwa.domain.vote.entity.place.PlaceVote;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
