+++ src/main/java/com/prography/yakgwa/domain/vote/repository/TimeVoteJpaRepository.java
@@ -1,6 +1,6 @@
 package com.prography.yakgwa.domain.vote.repository;
 
-import com.prography.yakgwa.domain.vote.entity.TimeVote;
+import com.prography.yakgwa.domain.vote.entity.time.TimeVote;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
